Astrometrie fonctionnelle Nouvelle version de mathlib ... tous les calculs se font dorénavant en double reel sur 64 bits TODO   La lecture du fichier de suivi pose probleme   PanelCourbe se deplace dans le buffer par saut de 4 octets   incompatible avec les doubles

+++ src/camera_mgr.cpp
@@ -513,7 +513,7 @@ void  Camera_mgr::tex2screen(int& x, int& y)
 //--------------------------------------------------------------------------------------------------------------------
 //
 //--------------------------------------------------------------------------------------------------------------------
-void  Camera_mgr::tex2screen(float& x,float& y)
+void  Camera_mgr::tex2screen(double& x,double& y)
 {
     if (pCurrent == NULL)                   return;
     pCurrent->getPanelPreview()->tex2screen(x, y);
@@ -529,7 +529,7 @@ void  Camera_mgr::screen2tex(int& x,int& y)
 //--------------------------------------------------------------------------------------------------------------------
 //
 //--------------------------------------------------------------------------------------------------------------------
-void  Camera_mgr::screen2tex(float& x,float& y)
+void  Camera_mgr::screen2tex(double& x,double& y)
 {
     if (pCurrent == NULL)                   return;
     pCurrent->getPanelPreview()->screen2tex(x, y);
